requirements.txt
biopython==1.77
matplotlib==3.2.1
numpy==1.18.4
pandas==1.0.3
tqdm==4.46.0
## The following requirements were added by pip freeze:
cycler==0.10.0
kiwisolver==1.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
